+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -171,6 +171,11 @@ New checks
 New check aliases
 ^^^^^^^^^^^^^^^^^
 
+- New alias :doc:`cert-ctr56-cpp
+  <clang-tidy/checks/cert/ctr56-cpp>` for
+  :doc:`bugprone-pointer-arithmetic-on-polymorphic-object
+  <clang-tidy/checks/bugprone/pointer-arithmetic-on-polymorphic-object>`.
+
 Changes in existing checks
 ^^^^^^^^^^^^^^^^^^^^^^^^^^